Move attack calculations into their own package
Attack generation is useful in move-gen, where this code originated,
but it can be used elsewhere. Like calculating pins & discovery checks.

Also move gen was cleaned up a little, with more work on defining the
API, but no implementation yet.  Test utilities were added as well, but
are unused until there is something to test.